refactor: split home page into sections
